+++ resolwe/flow/models/data.py
@@ -216,14 +216,14 @@ class DataQuerySet(BaseQuerySet, PermissionQuerySet):
             BackgroundTaskType.DUPLICATE_DATA, "Duplicate data", task_data, contributor
         )
 
-    def delete_background(self):
+    def delete_background(self, contributor):
         """Delete the ``Data`` objects in the background."""
         task_data = {
             "object_ids": list(self.values_list("pk", flat=True)),
-            "content_type_id": ContentType.objects.get_for_model(self).pk,
+            "content_type_id": ContentType.objects.get_for_model(self.model).pk,
         }
         return start_background_task(
-            BackgroundTaskType.DELETE, "Delete data", task_data, self.contributor
+            BackgroundTaskType.DELETE, "Delete data", task_data, contributor
         )
 
     @transaction.atomic